requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
marshmallow==3.21.3
marshmallow-sqlalchemy==1.1.0
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
flask-cors==6.0.1